expand_commands now in expander.c
