+++ src/logic/executor/expander.c
@@ -6,7 +6,7 @@
 /*   By: ismo <ismo@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/03/26 12:38:15 by ismo          #+#    #+#                 */
-/*   Updated: 2025/03/26 12:59:07 by ismo          ########   odam.nl         */
+/*   Updated: 2025/03/27 00:45:10 by ismo          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,4 +113,18 @@ void	expand_command(t_command *command)
 		// expand redir
 	}
 
+}
+
+void	expand_commands(t_list **commands)
+{
+	int	i;
+	t_list_node	*node;
+
+	i = 0;
+	node = get_node(i, *commands);
+	while (node)
+	{
+		expand_command((t_command *)node->value);
+		node = get_node(++i, *commands);
+	}
 }
