+++ src/logic/executor/expander.h
@@ -1,3 +1,16 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   expander.h                                         :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: ismo <ismo@student.codam.nl>                 +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2025/03/27 00:45:36 by ismo          #+#    #+#                 */
+/*   Updated: 2025/03/27 00:45:36 by ismo          ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef EXPANDER_H
 	void	expand_command(t_command *command);
+	void	expand_commands(t_list **commands);
 #endif
